Move the rt comparison chart
